requirements.txt
openai==1.40.6
openpyxl==3.1.2
pandas==2.2.2
plotly==5.23.0
python_docx==1.1.2
streamlit==1.37.1
